pyproject.toml
[project]
name = "eod-market-pipeline"
version = "0.1.0"
description = "End-of-day market data pipeline for Polygon.io ingestion, validation, and feature generation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.5.1",
    "fredapi>=0.5.2",
    "httpx>=0.28.1",
    "numpy>=2.4.4",
    "pandas>=3.0.1",
    "pandera>=0.30.1",
    "pyarrow>=23.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
markers = [
    "fred: tests for the FRED API",
    "massive: tests for the MASSIVE API",
]
